refactor(tui): extract template/ into component directory with types.rs and run.rs
Reorganizes the template component to the canonical complex-component
pattern:
- components/template/mod.rs: TUI form (renamed from create_config.rs)
- components/template/types.rs: TemplateParams, TemplateResult (serde)
- components/template/run.rs: execute() shared by JSON, CLI, and TUI

cmd.rs slimmed to JSON envelope types (JsonOutput, JsonError) plus
re-exports from the template component.
