+++ packages/keystone-tui/src/components/mod.rs
@@ -1,9 +1,9 @@
 pub mod build;
-pub mod create_config;
 pub mod deploy;
 pub mod first_boot;
 pub mod host_detail;
 pub mod hosts;
 pub mod install;
 pub mod iso;
+pub mod template;
 pub mod welcome;
